feat: allow different key price for change
We may want to use different key price for our side and their side. This is useful for bots that want to buy keys at a lower price than they sell them, even during other item trades.
For example, we may want to buy keys at 50 metal and sell keys at 60 metal. So if we sell a 5 ref item, but the buyer only has keys, then we want to only add 45 ref plus our item for their 1 key.
Similarly, if we buy a 5 ref item, but we only have keys then we want them to add 55 ref plus the item for our 1 key.